A simple and fast parser for RFC-3986 in C
It strictly follows the grammar given in the RFC, breaking the URI into its
component parts and delimiters.  It does not allocate or copy strings by
default.  There is some potential for improvement, since the are a couple
potentially inefficient edge-cases.  Specifically, the parser always tries to
get a userinfo first, which could force the parser to parse the same substring
twice, so if the host name is incredibly long this is inefficient.  Secondly,
IPv6 addresses have a lot of cases.  The are relatively short, but if you have
to try them all it does not gradually move forward, so must reparse things.