Avoid backtracking when userinfo is empty
It is very common for the userinfo to be missing in URIs.  Previously, the
parser would go forward until it couldn't anymore, and if it found an @ it would
make it a userinfo, but if @ was missing it would start over, possibly requiring
a complete reparse of the URI.  Now instead, if the @ is missing, it just makes
it into the host.  This works because the host grammar is almost a subset of the
userinfo.  There are two differences: IPv6 hosts start with a left bracket, but
that's okay because the userinfo will be empty.  Also, hosts can't include
colons, as they precede ports, while userinfos use them to precede passwords.