Parser for RFC 3966
This has a slight deviation from the strict spec, in that it allows flexible
parameter order.  To strictly check the order, define RFC_3966_CHECK_ORDER.

There is a small bug in parse_reserved.  The spec allows semicolons, but this
royally screws with everything and so for now they must be left out.