An enumeration of the URI schemes officially regestered by IANA
Many of these are provisional, many are permanent.  We don't distinguish between
these.

The comments give relevant information about the scheme, including any
description or RFC if relevant, as well as the original scheme in the case where
a character appears in the scheme that C cannot have in an identifier (where it
is replaced with an underscore).